changes in the code

+++ wbook-api/Infrastructure/Entities/Bundle.cs
@@ -8,8 +8,8 @@ namespace Infrastructure.Entities
         public int Id { get; set; }
         public string Title { get; set; } = null!;
         public string Description { get; set; } = null!;
-        public int Price { get; set; }
-        public int? Rrp { get; set; }
+        public double Price { get; set; }
+        public double? Rrp { get; set; }
         public string Image { get; set; } = null!;
         public string? Category { get; set; }
         public DateTime CreatedAt { get; set; }
