+++ wbook-api/Infrastructure/Entities/GameVariant.cs
@@ -9,8 +9,8 @@ namespace Infrastructure.Entities
         public int GameId { get; set; }
         public string Title { get; set; } = null!;
         public string Description { get; set; } = null!;
-        public int Price { get; set; }
-        public int? Rrp { get; set; }
+        public double Price { get; set; }
+        public double? Rrp { get; set; }
         public string Image { get; set; } = null!;
         public string? HoverImage { get; set; }
         public DateTime CreatedAt { get; set; }
